+++ tomcat/src/main/java/org/apache/coyote/http11/RequestMapping.java
@@ -1,8 +1,13 @@
-package org.apache.coyote.http11.controller;
+package org.apache.coyote.http11;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
+import org.apache.coyote.http11.controller.Controller;
+import org.apache.coyote.http11.controller.HomePageController;
+import org.apache.coyote.http11.controller.LoginController;
+import org.apache.coyote.http11.controller.RegisterController;
+import org.apache.coyote.http11.controller.StaticResourceController;
 import org.apache.coyote.http11.request.HttpRequest;
 
 public class RequestMapping {
